demo/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.16.2
Pillow==5.4.1
pkg-resources==0.0.0
protobuf==3.6.1
scipy==1.2.1
six==1.12.0
tensorboardX==1.6
torch==1.0.1.post2
torchvision==0.2.1
Werkzeug==0.14.1
